Replace fang with custom job queue (untested)
I'm not entirely sure if this works, but I had fun building it! The main
reason I needed to build something is that I want to be able to thread
in global handles to things (mostly the open DB transaction).

This may probably just converge onto the fang async stuff as I run into
all the same bugs. But maybe not?